pyproject.toml
[project]
name = "socket_basics"
version = "1.0.2"
description = "Socket Basics with integrated SAST, secret scanning, and container analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Socket.dev", email = "support@socket.dev"},
]
keywords = ["security", "sast", "secrets", "container", "scanning", "opengrep", "trivy", "trufflehog"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators", 
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Keep runtime deps minimal and explicit. Remove unused packages (mdutils, PyGithub)
    "requests>=2.31.0",
    "tabulate~=0.9.0",
    "light-s3-client~=0.0.30",
    "PyYAML>=6.0.0",
    "tomli; python_version < '3.11'",
    "socketdev>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.urls]
Homepage = "https://socket.dev"
Repository = "https://github.com/SocketDev/socket-basics"
Documentation = "https://github.com/SocketDev/socket-basics/blob/main/README.md"
"Bug Tracker" = "https://github.com/SocketDev/socket-basics/issues"

[project.scripts]
socket-basics = "socket_basics:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "socket_basics"]

[tool.hatch.build]
include = [
    "socket_basics/*.py",
    "socket_basics/rules/*.yml",
    "src/**/*.py",
]

[dependency-groups]
dev = []

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
